+++ core/XH.ts
@@ -8,7 +8,6 @@ import {
     HoistService,
     AppSpec,
     AppState,
-    createElement,
     Exception,
     ExceptionHandlerOptions,
     ExceptionHandler,
@@ -26,7 +25,7 @@ import {instanceManager} from './impl/InstanceManager';
 import {installServicesAsync} from './impl/InstallServices';
 import {Icon} from '@xh/hoist/icon';
 import {action} from '@xh/hoist/mobx';
-import {never, wait} from '@xh/hoist/promise';
+import {never} from '@xh/hoist/promise';
 import {
     AlertBannerService,
     AutoRefreshService,
@@ -46,26 +45,15 @@ import {
     WebSocketService,
     FetchOptions
 } from '@xh/hoist/svc';
-import {MINUTES} from '@xh/hoist/utils/datetime';
-import {checkMinVersion, throwIf} from '@xh/hoist/utils/js';
-import {camelCase, compact, flatten, isString, uniqueId} from 'lodash';
-import {createRoot} from 'react-dom/client';
+import {camelCase, flatten, isString, uniqueId} from 'lodash';
 import {AppContainerModel} from '../appcontainer/AppContainerModel';
 import {ToastModel} from '../appcontainer/ToastModel';
 import {BannerModel} from '../appcontainer/BannerModel';
 import '../styles/XH.scss';
 import {ModelSelector, HoistModel, RefreshContextModel} from './model';
-import {
-    HoistAppModel,
-    RouterModel,
-    BannerSpec,
-    ToastSpec,
-    MessageSpec,
-    HoistUser,
-    TaskObserver
-} from './';
+import {HoistAppModel, BannerSpec, ToastSpec, MessageSpec, HoistUser, TaskObserver} from './';
 
-const MIN_HOIST_CORE_VERSION = '16.0';
+export const MIN_HOIST_CORE_VERSION = '16.0';
 
 declare const xhAppCode: string;
 declare const xhAppName: string;
@@ -88,7 +76,6 @@ export class XHApi {
     // Core Delegates
     //----------------
     appContainerModel: AppContainerModel = new AppContainerModel();
-    routerModel: RouterModel = new RouterModel();
     exceptionHandler: ExceptionHandler = new ExceptionHandler();
 
     //----------------------------------------------------------------------------------------------
@@ -245,38 +232,34 @@ export class XHApi {
         return this.acm.pageStateModel.state;
     }
 
-    /** current lifecycle state of the application. */
+    /** Current lifecycle state of the application. */
     get appState(): AppState {
         return this.acm.appStateModel.state;
     }
 
-    /** milliseconds timestamp at moment user activity / interaction was last detected. */
+    /** Milliseconds timestamp at moment user activity / interaction was last detected. */
     get lastActivityMs(): number {
         return this.acm.appStateModel.lastActivityMs;
     }
 
-    /** true if application initialized and running (observable). */
+    /** True if application initialized and running (observable). */
     get appIsRunning(): boolean {
         return this.appState === 'RUNNING';
     }
 
     /** Root level application model. */
-    appModel: HoistAppModel = null;
+    get appModel(): HoistAppModel {
+        return this.acm.appModel;
+    }
 
     /** Specifications for this application, provided in call to `XH.renderApp()`. */
-    appSpec: AppSpec = null;
+    get appSpec(): AppSpec {
+        return this.acm.appSpec;
+    }
 
     /** Main entry point. Initialize and render application code. */
     renderApp<T extends HoistAppModel>(appSpec: AppSpec<T>) {
-        // Remove the pre-load exception handler installed by preflight.js
-        window.onerror = null;
-        const spinner = document.getElementById('xh-preload-spinner');
-        if (spinner) spinner.style.display = 'none';
-        this.appSpec = appSpec instanceof AppSpec ? appSpec : new AppSpec(appSpec);
-
-        const root = createRoot(document.getElementById('xh-root')),
-            rootView = createElement(appSpec.containerClass, {model: this.appContainerModel});
-        root.render(rootView);
+        this.acm.renderApp(appSpec);
     }
 
     /**
@@ -369,12 +352,12 @@ export class XHApi {
     //------------------------
     // Sizing Mode Support
     //------------------------
-    setSizingMode(sizingMode: SizingMode) {
-        return this.acm.sizingModeModel.setSizingMode(sizingMode);
-    }
-
     get sizingMode(): SizingMode {
         return this.acm.sizingModeModel.sizingMode;
+    }
+
+    setSizingMode(sizingMode: SizingMode) {
+        return this.acm.sizingModeModel.setSizingMode(sizingMode);
     }
 
     //------------------------
@@ -398,6 +381,10 @@ export class XHApi {
     //-------------------------
     // Routing support
     //-------------------------
+    get routerModel() {
+        return this.acm.routerModel;
+    }
+
     /**
      * Underlying Router5 Router object implementing the routing state.
      * Applications should use this property to directly access the Router5 API.
@@ -661,169 +648,9 @@ export class XHApi {
         return uniqueId('xh-id-');
     }
 
-    //---------------------------------
-    // Framework Methods
-    //---------------------------------
-    private initCalled = false;
-    /**
-     * Called when application container first mounted in order to trigger initial
-     * authentication and initialization of framework and application.
-     * @internal
-     */
-    async initAsync() {
-        // Avoid multiple calls, which can occur if AppContainer remounted.
-        if (this.initCalled) return;
-        this.initCalled = true;
-
-        const {appSpec, isMobileApp, isPhone, isTablet, isDesktop} = this;
-
-        // Add xh css classes to power Hoist CSS selectors.
-        document.body.classList.add(
-            ...compact([
-                'xh-app',
-                isMobileApp ? 'xh-mobile' : 'xh-standard',
-                isDesktop ? 'xh-desktop' : null,
-                isPhone ? 'xh-phone' : null,
-                isTablet ? 'xh-tablet' : null
-            ])
-        );
-
-        // Disable browser context menu on long-press, used to show (app) context menus and as an
-        // alternate gesture for tree grid drill-own.
-        if (isMobileApp) {
-            window.addEventListener('contextmenu', e => e.preventDefault(), {capture: true});
-        }
-
-        try {
-            await this.installServicesAsync(FetchService);
-            this.setAppState('PRE_AUTH');
-
-            // consult (optional) pre-auth init for app
-            const modelClass: any = this.appSpec.modelClass;
-            await modelClass.preAuthAsync();
-
-            // Check if user has already been authenticated (prior login, OAuth, SSO)...
-            const userIsAuthenticated = await this.getAuthStatusFromServerAsync();
-
-            // ...if not, throw in SSO mode (unexpected error case) or trigger a login prompt.
-            if (!userIsAuthenticated) {
-                throwIf(
-                    appSpec.isSSO,
-                    'Unable to complete required authentication (SSO/Oauth failure).'
-                );
-                this.setAppState('LOGIN_REQUIRED');
-                return;
-            }
-
-            // ...if so, continue with initialization.
-            await this.completeInitAsync();
-        } catch (e) {
-            this.setAppState('LOAD_FAILED');
-            this.handleException(e, {requireReload: true});
-        }
-    }
-
-    /**
-     * Complete initialization. Called after the client has confirmed that the user is generally
-     * authenticated and known to the server (regardless of application roles at this point).
-     * @internal
-     */
-    @action
-    async completeInitAsync() {
-        try {
-            // Install identity service and confirm access
-            await this.installServicesAsync(IdentityService);
-            if (!this.identityService.checkAccess()) {
-                this.setAppState('ACCESS_DENIED');
-                return;
-            }
-
-            // Complete initialization process
-            this.setAppState('INITIALIZING');
-            await this.installServicesAsync(ConfigService, LocalStorageService);
-            await this.installServicesAsync(TrackService);
-            await this.installServicesAsync(EnvironmentService, PrefService, JsonBlobService);
-
-            // Confirm hoist-core version after environment service loaded
-            const hcVersion = XH.environmentService.get('hoistCoreVersion');
-            if (!checkMinVersion(hcVersion, MIN_HOIST_CORE_VERSION)) {
-                throw XH.exception(`
-                    This version of Hoist React requires the server to run Hoist Core
-                    v${MIN_HOIST_CORE_VERSION} or greater. Version ${hcVersion} detected.
-                `);
-            }
-
-            await this.installServicesAsync(
-                AlertBannerService,
-                AutoRefreshService,
-                ChangelogService,
-                IdleService,
-                InspectorService,
-                GridAutosizeService,
-                GridExportService,
-                WebSocketService
-            );
-            this.acm.init();
-
-            this.setDocTitle();
-
-            // Delay to workaround hot-reload styling issues in dev.
-            await wait(XH.isDevelopmentMode ? 300 : 1);
-
-            const modelClass: any = this.appSpec.modelClass;
-            this.appModel = modelClass.instance = new modelClass();
-            await this.appModel.initAsync();
-            this.startRouter();
-            this.startOptionsDialog();
-            this.setAppState('RUNNING');
-        } catch (e) {
-            this.setAppState('LOAD_FAILED');
-            this.handleException(e, {requireReload: true});
-        }
-    }
-
-    //------------------------
-    // Implementation
-    //------------------------
-    private setAppState(nextState: AppState) {
-        this.acm.appStateModel.setAppState(nextState);
-    }
-
-    private setDocTitle() {
-        const env = XH.getEnv('appEnvironment'),
-            {clientAppName} = this.appSpec;
-        document.title = env === 'Production' ? clientAppName : `${clientAppName} (${env})`;
-    }
-
-    private async getAuthStatusFromServerAsync(): Promise<boolean> {
-        return await this.fetchService
-            .fetchJson({
-                url: 'xh/authStatus',
-                timeout: 3 * MINUTES // Accommodate delay for user at a credentials prompt
-            })
-            .then(r => r.authenticated)
-            .catch(e => {
-                // 401s normal / expected for non-SSO apps when user not yet logged in.
-                if (e.httpStatus === 401) return false;
-                // Other exceptions indicate e.g. connectivity issue, server down - raise to user.
-                throw e;
-            });
-    }
-
-    private startRouter() {
-        this.routerModel.addRoutes(this.appModel.getRoutes());
-        this.router.start();
-    }
-
-    private startOptionsDialog() {
-        this.acm.optionsDialogModel.setOptions(this.appModel.getAppOptions());
-    }
-
     private get acm(): AppContainerModel {
         return this.appContainerModel;
     }
-
-    private parseAppSpec() {}
 }
 
 /** app-wide singleton instance. */
